requirements.txt
python-telegram-bot
google-genai
google-generativeai